Fix volume read status
Previously, the volume reported by Home Assistant for my LG TV was always "NaN".
This is due to the fact that the payload returned by the library contains "volume",
but wrapped in an outer structure of "volumeStatus". This PR fixes this and now
Home Assistant correctly reports the current value of volume.

Tested by running modified version of code in Home Assistant.